pyproject.toml
[project]
name = "rag-bonus-pack"
version = "0.1.0"
description = "RAG Chatbot: Prueba técnica de MasOrange"
authors = [{ name = "Celia Patricio" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "tiktoken>=0.7.0",
  "openai>=1.30.0",
  "pyyaml>=6.0",
  "pytest>=8.0.0",
  "python-dotenv>=0.21.0",
]


[tool.uv]
# Sin configuraciones especiales: uv usa pyproject por defecto


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"


[tool.ruff]
line-length = 100
